@syrusakbary 2019-08-13
Previous timing:

```
wasmer run --backend=llvm --disable-cache --enable-simd   18.95s user 0.08s system 99% cpu 19.114 total
```

With the changes from this PR:
```
/Users/syrusakbary/Development/wasmer/target/release/wasmer run --backend=llv  14.21s user 0.03s system 99% cpu 14.263 total
```