@penzn 2019-08-16
Hi, I have reported the regression originally in wasmerio/c-wasm-simd128-example#1. If I reset `master` to commit right before 11f66d287b7e9ee81f49315c44e45f0089acf26a then I see the timings from the [article][a], but timings from the branch in this PR is still similar to `master`. Is that me or are you seeing that too?

[a]: https://medium.com/wasmer/webassembly-and-simd-13badb9bf1a8