@nlewycky 2019-12-02
This PR is obsoleted by PR #883 and PR #934.