@nlewycky 2019-12-02
@penzn Yes, we noticed the same thing. It's a bit of a pain because LLVM doesn't say a lot about what happens with NaN values on operations, even as simple as an `fadd` instruction. It happened to work before, because it lowers to an addition that happens to preserve NaNs. Unfortunately, while the work I did to try to bring performance back did help on many examples, it hasn't helped to improve the c-wasm-simd128-example when built with SIMD. The primary difference there remains whether the LLVM autovectorizer manages to vectorize the `fdiv` instruction. That can be affected by many things, we've also changed the LLVM pass list to improve optimizations which should improve the chance that we get the vectorized fdiv in the simd example, but unfortunately we still don't. I'd still like to figure out how to get guaranteed correctness while preserving the performance, especially since it should be possible given that the x86-64 instructions give us what we need.