@penzn 2019-12-03
@nlewycky I just rebuilt the benchmark with `-munimplemented-simd128`, which leads to vector `fdiv` in the wasm module, and performance is the same as with scalar `fdiv`. It also looks relatively close to native (given that I keep caching on). Can you elaborate a bit on what you are seeing, I'd be happy to help.

Maybe we should open an issue for this.

CC @topperc